pyproject.toml
[tool.poetry]
name = "dados_ifood"
version = "0.1.0"
description = ""
authors = ["Rodrigoneal <rodrigho2006@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
selenium-tools = {git = "https://github.com/digital-path/selenium_tools"}
pandas = "^1.4.3"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^22.6.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
